requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.6
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.10
whitenoise==6.2.0
